editor/libeditor/tests/test_dom_input_event_on_htmleditor.html

<html>
<head>
  <title>Test for input event of text editor</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css"
          href="/tests/SimpleTest/test.css" />
</head>
<body>
<div id="display">
  <iframe id="editor1" srcdoc="<html><body contenteditable id='eventTarget'></body></html>"></iframe>
  <iframe id="editor2" srcdoc="<html contenteditable id='eventTarget'><body></body></html>"></iframe>
  <iframe id="editor3" srcdoc="<html><body><div contenteditable id='eventTarget'></div></body></html>"></iframe>
  <iframe id="editor4" srcdoc="<html contenteditable id='eventTarget'><body><div contenteditable></div></body></html>"></iframe>
  <iframe id="editor5" srcdoc="<html><body id='eventTarget'></body><script>document.designMode='on';</script></html>"></iframe>
</div>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(runTests, window);

const kIsWin = navigator.platform.indexOf("Win") == 0;
const kIsMac = navigator.platform.indexOf("Mac") == 0;

function runTests() {
  function doTests(aDocument, aWindow, aDescription) {
    aDescription += ": ";
    aWindow.focus();

    var body = aDocument.body;
    var selection = aWindow.getSelection();

    var eventTarget = aDocument.getElementById("eventTarget");
    // The event target must be focusable because it's the editing host.
    eventTarget.focus();

    var editTarget = aDocument.getElementById("editTarget");
    if (!editTarget) {
      editTarget = eventTarget;
    }

    // Root element never can be edit target.  If the editTarget is the root
    // element, replace with its body.
    if (editTarget == aDocument.documentElement) {
      editTarget = body;
    }

    editTarget.innerHTML = "";

    // If the editTarget isn't its editing host, move caret to the start of it.
    if (eventTarget != editTarget) {
      aDocument.getSelection().collapse(editTarget, 0);
    }

    var inputEvent = null;

    var handler = function(aEvent) {
      is(aEvent.target, eventTarget,
         "input event is fired on unexpected element: " + aEvent.target.tagName);
      ok(aEvent instanceof InputEvent,
         "input event should be dispatched with InputEvent interface");
      ok(!aEvent.cancelable, "input event must not be cancelable");
      ok(aEvent.bubbles, "input event must be bubbles");
      let duration = Math.abs(window.performance.now() - aEvent.timeStamp);
      ok(duration < 30 * 1000,
         "perhaps, timestamp wasn't set correctly :" + aEvent.timeStamp +
         " (expected it to be within 30s of the current time but it " +
         "differed by " + duration + "ms)");
      inputEvent = aEvent;
    };

    aWindow.addEventListener("input", handler, true);

    inputEvent = null;
    synthesizeKey("a", { }, aWindow);
    is(editTarget.innerHTML, "a", aDescription + "wrong element was edited");
    ok(inputEvent, aDescription + "input event wasn't fired by 'a' key");
    ok(inputEvent.isTrusted, aDescription + "input event by 'a' key wasn't trusted event");
    is(inputEvent.inputType, "insertText",
       aDescription + 'inputType should be "insertText" when typing "a"');
    is(inputEvent.data, "a",
       aDescription + 'data should be "a" when typing "a"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when typing "a"');

    inputEvent = null;
    synthesizeKey("KEY_Backspace", { }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by BackSpace key");
    ok(inputEvent.isTrusted, aDescription + "input event by BackSpace key wasn't trusted event");
    is(inputEvent.inputType, "deleteContentBackward",
       aDescription + 'inputType should be "deleteContentBackward" when pressing "Backspace" with collapsed selection');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing "Backspace" with collapsed selection');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing "Backspace" with collapsed selection');

    inputEvent = null;
    synthesizeKey("B", { shiftKey: true }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by 'B' key");
    ok(inputEvent.isTrusted, aDescription + "input event by 'B' key wasn't trusted event");
    is(inputEvent.inputType, "insertText",
       aDescription + 'inputType should be "insertText" when typing "B"');
    is(inputEvent.data, "B",
       aDescription + 'data should be "B" when typing "B"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when typing "B"');

    inputEvent = null;
    synthesizeKey("KEY_Enter", { }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by Enter key");
    ok(inputEvent.isTrusted, aDescription + "input event by Enter key wasn't trusted event");
    is(inputEvent.inputType, "insertParagraph",
       aDescription + 'inputType should be "insertParagraph" when pressing "Enter"');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing "Enter"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing "Enter"');

    inputEvent = null;
    synthesizeKey("C", { shiftKey: true }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by 'C' key");
    ok(inputEvent.isTrusted, aDescription + "input event by 'C' key wasn't trusted event");
    is(inputEvent.inputType, "insertText",
       aDescription + 'inputType should be "insertText" when typing "C"');
    is(inputEvent.data, "C",
       aDescription + 'data should be "C" when typing "C"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when typing "C"');

    inputEvent = null;
    synthesizeKey("KEY_Enter", { }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by Enter key (again)");
    ok(inputEvent.isTrusted, aDescription + "input event by Enter key (again) wasn't trusted event");
    is(inputEvent.inputType, "insertParagraph",
       aDescription + 'inputType should be "insertParagraph" when pressing "Enter" again');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing "Enter" again');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing "Enter" again');

    inputEvent = null;
    editTarget.innerHTML = "foo-bar";
    ok(!inputEvent, aDescription + "input event was fired by setting value");

    inputEvent = null;
    editTarget.innerHTML = "";
    ok(!inputEvent, aDescription + "input event was fired by setting empty value");

    inputEvent = null;
    synthesizeKey(" ", { }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by Space key");
    ok(inputEvent.isTrusted, aDescription + "input event by Space key wasn't trusted event");
    is(inputEvent.inputType, "insertText",
       aDescription + 'inputType should be "insertText" when typing " "');
    is(inputEvent.data, " ",
       aDescription + 'data should be " " when typing " "');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when typing " "');

    inputEvent = null;
    synthesizeKey("KEY_Delete", { }, aWindow);
    ok(!inputEvent, aDescription + "input event was fired by Delete key at the end");

    inputEvent = null;
    synthesizeKey("KEY_ArrowLeft", { }, aWindow);
    ok(!inputEvent, aDescription + "input event was fired by Left key");

    inputEvent = null;
    synthesizeKey("KEY_Delete", { }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by Delete key at the start");
    ok(inputEvent.isTrusted, aDescription + "input event by Delete key wasn't trusted event");
    is(inputEvent.inputType, "deleteContentForward",
       aDescription + 'inputType should be "deleteContentForward" when pressing "Delete" with collapsed selection');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing "Delete" with collapsed selection');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing "Delete" with collapsed selection');

    inputEvent = null;
    synthesizeKey("z", { accelKey: true }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by Undo");
    ok(inputEvent.isTrusted, aDescription + "input event by Undo wasn't trusted event");
    is(inputEvent.inputType, "historyUndo",
       aDescription + 'inputType should be "historyUndo" when doing "Undo"');
    is(inputEvent.data, null,
       aDescription + 'data should be null when doing "Undo"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when doing "Undo"');

    inputEvent = null;
    synthesizeKey("z", { accelKey: true, shiftKey: true }, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by Redo");
    ok(inputEvent.isTrusted, aDescription + "input event by Redo wasn't trusted event");
    is(inputEvent.inputType, "historyRedo",
       aDescription + 'inputType should be "historyRedo" when doing "Redo"');
    is(inputEvent.data, null,
       aDescription + 'data should be null when doing "Redo"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when doing "Redo"');

    inputEvent = null;
    synthesizeKey("KEY_Enter", {shiftKey: true}, aWindow);
    ok(inputEvent, aDescription + "input event wasn't fired by Shift + Enter key");
    ok(inputEvent.isTrusted, aDescription + "input event by Shift + Enter key wasn't trusted event");
    is(inputEvent.inputType, "insertLineBreak",
       aDescription + 'inputType should be "insertLineBreak" when pressing Shift + "Enter"');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing Shift + "Enter"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing Shift + "Enter"');

    // Backspace/Delete with non-collapsed selection.
    editTarget.innerHTML = "a";
    editTarget.focus();
    selection.selectAllChildren(editTarget);
    inputEvent = null;
    synthesizeKey("KEY_Backspace", {}, aWindow);
    ok(inputEvent,
       aDescription + 'input event should be fired by pressing "Backspace" with non-collapsed selection');
    ok(inputEvent.isTrusted,
       aDescription + 'input event should be trusted when pressing "Backspace" with non-collapsed selection');
    is(inputEvent.inputType, "deleteContentBackward",
       aDescription + 'inputType should be "deleteContentBackward" when pressing "Backspace" with non-collapsed selection');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing "Backspace" with non-collapsed selection');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing "Backspace" with non-collapsed selection');

    editTarget.innerHTML = "a";
    editTarget.focus();
    selection.selectAllChildren(editTarget);
    inputEvent = null;
    synthesizeKey("KEY_Delete", {}, aWindow);
    ok(inputEvent,
       aDescription + 'input event should be fired by pressing "Delete" with non-collapsed selection');
    ok(inputEvent.isTrusted,
       aDescription + 'input event should be trusted when pressing "Delete" with non-collapsed selection');
    is(inputEvent.inputType, "deleteContentForward",
       aDescription + 'inputType should be "deleteContentBackward" when Delete "Backspace" with non-collapsed selection');
    is(inputEvent.data, null,
       aDescription + 'data should be null when Delete "Backspace" with non-collapsed selection');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when Delete "Backspace" with non-collapsed selection');

    // Delete to previous/next word boundary with collapsed selection.
    editTarget.innerHTML = "a";
    editTarget.focus();
    selection.selectAllChildren(editTarget);
    selection.collapseToEnd();
    inputEvent = null;
    SpecialPowers.doCommand(aWindow, "cmd_deleteWordBackward");
    ok(inputEvent,
       aDescription + "input event should be fired by deleting to previous word boundary with collapsed selection");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when deleting to previous word boundary with collapsed selection");
    is(inputEvent.inputType, "deleteWordBackward",
       aDescription + 'inputType should be "deleteWordBackward" when deleting to previous word boundary with collapsed selection');
    is(inputEvent.data, null,
       aDescription + "data should be null when deleting to previous word boundary with collapsed selection");
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when deleting to previous word boundary with collapsed selection");

    editTarget.innerHTML = "a";
    editTarget.focus();
    selection.selectAllChildren(editTarget);
    selection.collapseToStart();
    inputEvent = null;
    SpecialPowers.doCommand(aWindow, "cmd_deleteWordForward");
    ok(inputEvent,
       aDescription + "input event should be fired by deleting to next word boundary with collapsed selection");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when deleting to next word boundary with collapsed selection");
    is(inputEvent.inputType, "deleteWordForward",
       aDescription + 'inputType should be "deleteWordForward" when deleting to next word boundary with collapsed selection');
    is(inputEvent.data, null,
       aDescription + "data should be null when deleting to next word boundary with collapsed selection");
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when deleting to next word boundary with collapsed selection");

    // Delete to previous/next word boundary with non-collapsed selection.
    editTarget.innerHTML = "abc";
    editTarget.focus();
    selection.setBaseAndExtent(editTarget.firstChild, 1, editTarget.firstChild, 2);
    inputEvent = null;
    SpecialPowers.doCommand(aWindow, "cmd_deleteWordBackward");
    ok(inputEvent,
       aDescription + "input event should be fired by deleting to previous word boundary with non-collapsed selection");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when deleting to previous word boundary with non-collapsed selection");
    if (kIsWin) {
      // Only on Windows, we collapse selection to start before handling this command.
      is(inputEvent.inputType, "deleteWordBackward",
         aDescription + 'inputType should be "deleteWordBackward" when deleting to previous word boundary with non-collapsed selection');
    } else {
      is(inputEvent.inputType, "deleteContentBackward",
         aDescription + 'inputType should be "deleteContentBackward" when deleting to previous word boundary with non-collapsed selection');
    }
    is(inputEvent.data, null,
       aDescription + "data should be null when deleting to previous word boundary with non-collapsed selection");
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when deleting to previous word boundary with non-collapsed selection");

    editTarget.innerHTML = "abc";
    editTarget.focus();
    selection.setBaseAndExtent(editTarget.firstChild, 1, editTarget.firstChild, 2);
    inputEvent = null;
    SpecialPowers.doCommand(aWindow, "cmd_deleteWordForward");
    ok(inputEvent,
       aDescription + "input event should be fired by deleting to next word boundary with non-collapsed selection");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when deleting to next word boundary with non-collapsed selection");
    if (kIsWin) {
      // Only on Windows, we collapse selection to start before handling this command.
      is(inputEvent.inputType, "deleteWordForward",
         aDescription + 'inputType should be "deleteWordForward" when deleting to next word boundary with non-collapsed selection');
    } else {
      is(inputEvent.inputType, "deleteContentForward",
         aDescription + 'inputType should be "deleteContentForward" when deleting to next word boundary with non-collapsed selection');
    }
    is(inputEvent.data, null,
       aDescription + "data should be null when deleting to next word boundary with non-collapsed selection");
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when deleting to next word boundary with non-collapsed selection");

    // Delete to previous/next visual line boundary with collapsed selection.
    editTarget.innerHTML = "a";
    editTarget.focus();
    selection.selectAllChildren(editTarget);
    selection.collapseToEnd();
    inputEvent = null;
    SpecialPowers.doCommand(aWindow, "cmd_deleteToBeginningOfLine");
    ok(inputEvent,
       aDescription + "input event should be fired by deleting to previous visual line boundary with collapsed selection");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when deleting to previous visual line boundary with collapsed selection");
    is(inputEvent.inputType, "deleteSoftLineBackward",
       aDescription + 'inputType should be "deleteSoftLineBackward" when deleting to previous visual line boundary with collapsed selection');
    is(inputEvent.data, null,
       aDescription + "data should be null when deleting to previous visual line boundary with collapsed selection");
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when deleting to previous visual line boundary with collapsed selection");

    editTarget.innerHTML = "a";
    editTarget.focus();
    selection.selectAllChildren(editTarget);
    selection.collapseToStart();
    inputEvent = null;
    SpecialPowers.doCommand(aWindow, "cmd_deleteToEndOfLine");
    ok(inputEvent,
       aDescription + "input event should be fired by deleting to next visual line boundary with collapsed selection");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when deleting to next visual line boundary with collapsed selection");
    is(inputEvent.inputType, "deleteSoftLineForward",
       aDescription + 'inputType should be "deleteSoftLineForward" when deleting to visual line boundary with collapsed selection');
    is(inputEvent.data, null,
       aDescription + "data should be null when deleting to visual line boundary with collapsed selection");
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when deleting to visual line boundary with collapsed selection");

    // Delete to previous/next visual line boundary with non-collapsed selection.
    editTarget.innerHTML = "abc";
    editTarget.focus();
    selection.setBaseAndExtent(editTarget.firstChild, 1, editTarget.firstChild, 2);
    inputEvent = null;
    SpecialPowers.doCommand(aWindow, "cmd_deleteToBeginningOfLine");
    ok(inputEvent,
       aDescription + "input event should be fired by deleting to previous visual line boundary with non-collapsed selection");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when deleting to previous visual line boundary with non-collapsed selection");
    if (kIsWin) {
      // Only on Windows, we collapse selection to start before handling this command.
      is(inputEvent.inputType, "deleteSoftLineBackward",
         aDescription + 'inputType should be "deleteSoftLineBackward" when deleting to next visual line boundary with non-collapsed selection');
    } else {
      is(inputEvent.inputType, "deleteContentBackward",
         aDescription + 'inputType should be "deleteContentBackward" when deleting to previous visual line boundary with non-collapsed selection');
    }
    is(inputEvent.data, null,
       aDescription + "data should be null when deleting to previous visual line boundary with non-collapsed selection");
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when deleting to previous visual line boundary with non-collapsed selection");

    editTarget.innerHTML = "abc";
    editTarget.focus();
    selection.setBaseAndExtent(editTarget.firstChild, 1, editTarget.firstChild, 2);
    inputEvent = null;
    SpecialPowers.doCommand(aWindow, "cmd_deleteToEndOfLine");
    ok(inputEvent,
       aDescription + "input event should be fired by deleting to next visual line boundary with non-collapsed selection");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when deleting to next visual line boundary with non-collapsed selection");
    if (kIsWin) {
      // Only on Windows, we collapse selection to start before handling this command.
      is(inputEvent.inputType, "deleteSoftLineForward",
         aDescription + 'inputType should be "deleteSoftLineForward" when deleting to next visual line boundary with non-collapsed selection');
    } else {
      is(inputEvent.inputType, "deleteContentForward",
         aDescription + 'inputType should be "deleteContentForward" when deleting to next visual line boundary with non-collapsed selection');
    }
    is(inputEvent.data, null,
       aDescription + "data should be null when deleting to next visual line boundary with non-collapsed selection");
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when deleting to next visual line boundary with non-collapsed selection");

    // Toggling text direction
    editTarget.focus();
    inputEvent = null;
    SpecialPowers.doCommand(window, "cmd_switchTextDirection");
    ok(inputEvent,
       aDescription + "input event should be fired by dispatching cmd_switchTextDirection command #1");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when dispatching cmd_switchTextDirection command #1");
    is(inputEvent.inputType, "formatSetBlockTextDirection",
       aDescription + 'inputType should be "formatSetBlockTextDirection" when dispatching cmd_switchTextDirection command #1');
    is(inputEvent.data, "rtl",
       aDescription + 'data should be "rtl" when dispatching cmd_switchTextDirection command #1');
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when dispatching cmd_switchTextDirection command #1");

    inputEvent = null;
    SpecialPowers.doCommand(window, "cmd_switchTextDirection");
    ok(inputEvent,
       aDescription + "input event should be fired by dispatching cmd_switchTextDirection command #2");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when dispatching cmd_switchTextDirection command #2");
    is(inputEvent.inputType, "formatSetBlockTextDirection",
       aDescription + 'inputType should be "formatSetBlockTextDirection" when dispatching cmd_switchTextDirection command #2');
    is(inputEvent.data, "ltr",
       aDescription + 'data should be "ltr" when dispatching cmd_switchTextDirection command #2');
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when dispatching cmd_switchTextDirection command #2");

    // Inserting link
    editTarget.innerHTML = "link";
    editTarget.focus();
    selection.selectAllChildren(editTarget);
    inputEvent = null;
    aDocument.execCommand("createLink", false, "https://example.com/foo/bar.html");
    ok(inputEvent,
       aDescription + 'input event should be fired by execCommand("createLink", false, "https://example.com/foo/bar.html")');
    ok(inputEvent.isTrusted,
       aDescription + 'input event should be trusted when execCommand("createLink", false, "https://example.com/foo/bar.html")');
    is(inputEvent.inputType, "insertLink",
       aDescription + 'inputType should be "insertLink" when execCommand("createLink", false, "https://example.com/foo/bar.html")');
    is(inputEvent.data, "https://example.com/foo/bar.html",
       aDescription + 'data should be "https://example.com/foo/bar.html" when execCommand("createLink", false, "https://example.com/foo/bar.html")');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when execCommand("createLink", false, "https://example.com/foo/bar.html")');

    selection.selectAllChildren(editTarget);
    aDocument.execCommand("createLink", false, "foo/bar.html");
    ok(inputEvent,
       aDescription + 'input event should be fired by execCommand("createLink", false, "foo/bar.html")');
    ok(inputEvent.isTrusted,
       aDescription + 'input event should be trusted when execCommand("createLink", false, "foo/bar.html")');
    is(inputEvent.inputType, "insertLink",
       aDescription + 'inputType should be "insertLink" when execCommand("createLink", false, "foo/bar.html")');
    is(inputEvent.data, "foo/bar.html",
       aDescription + 'data should be "foo/bar.html" when execCommand("createLink", false, "foo/bar.html")');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when execCommand("createLink", false, "foo/bar.html")');

    aWindow.removeEventListener("input", handler, true);
  }

  doTests(document.getElementById("editor1").contentDocument,
          document.getElementById("editor1").contentWindow,
          "Editor1, body has contenteditable attribute");
  doTests(document.getElementById("editor2").contentDocument,
          document.getElementById("editor2").contentWindow,
          "Editor2, html has contenteditable attribute");
  doTests(document.getElementById("editor3").contentDocument,
          document.getElementById("editor3").contentWindow,
          "Editor3, div has contenteditable attribute");
  doTests(document.getElementById("editor4").contentDocument,
          document.getElementById("editor4").contentWindow,
          "Editor4, html and div has contenteditable attribute");
  doTests(document.getElementById("editor5").contentDocument,
          document.getElementById("editor5").contentWindow,
          "Editor5, html and div has contenteditable attribute");

  SimpleTest.finish();
}

</script>
</body>

</html>
